api/dev-requirements.txt
black==23.9.1
pytest==7.4.2
pytest-cov==4.1.0
PyHamcrest==2.0.4
httpx==0.25.0
